refactor: #893 Option 1 — split test_crud_operations_unit.py into 5 per-module files
Dissolve the monolithic tests/unit/database/test_crud_operations_unit.py
(3564 lines, 31 interleaved test classes) into 5 per-module test files
aligned with the crud_* modules they exercise:

- test_crud_events_unit.py        (6 classes,  33 tests)
- test_crud_markets_unit.py       (3 classes,   9 tests)
- test_crud_teams_unit.py         (7 classes,  28 tests)  # teams + venues + team_rankings
- test_crud_game_states_unit.py  (10 classes,  60 tests)  # game_states + games proper
- test_crud_schedulers_unit.py    (5 classes,  14 tests)

Total: 31 classes, 144 unique tests, 149 collected (5 parametrized expansions).
Zero test loss vs. the original file (verified by pytest --collect-only on both).

Pattern 74 rename-over-rewrite: class bodies, method names, parametrize
decorators, and assertions are preserved verbatim. Only the module-level
imports are trimmed per-file to the names each group actually uses. No
source code or fixtures are modified.

Closes 5 of 7 Class A #887 test-type-coverage audit gaps atomically:
audit_test_type_coverage.py now reports PASS for database/crud_events,
crud_markets, crud_teams, crud_game_states, and crud_schedulers.
The remaining 2 Class A gaps (crud_elo experimental, temporal_alignment_writer
business) are tracked separately.
